student_management_system/requirements.txt
Django==4.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
stripe==7.1.0
python-dotenv==1.0.0
Pillow==10.1.0
django-environ==0.11.2
django-cors-headers==4.3.1
whitenoise==6.6.0
dj-database-url==2.1.0
sqlparse==0.4.4
gunicorn==21.2.0
qrcode==7.4.2
asgiref==3.7.2
channels==4.0.0
daphne==4.0.0
channels_redis==4.0.0
